% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom.R
\name{create_custom_fd}
\alias{create_custom_fd}
\title{Create a custom feed ingredient}
\usage{
create_custom_fd(template_fd_name, new_fd_name, modifications)
}
\arguments{
\item{template_fd_name}{The Fd_Name of the template feed to use as a base for modifications.}

\item{new_fd_name}{The name of the new custom feed.}

\item{modifications}{A list of modifications to apply to the template feed.}
}
\value{
A dataframe representing the new custom feed with applied modifications.
}
\description{
Some users may wish to modify the feed library by adding custom ingredients. This function assumes that new feeds are modified from existing feeds in the default library. If preferred, it is also possible to modify a feed library in Excel or a csv editor rather than in R.
}
\examples{

create_custom_fd("Corn silage, typical", "My_Custom_CornSilage",
 list("Fd_DM" = 36.0, "Fd_CP" = 7.0))

}
