% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom.R
\name{create_custom_library}
\alias{create_custom_library}
\title{Create a custom feed library}
\usage{
create_custom_library(base_library = "default_library", custom_fd_list)
}
\arguments{
\item{base_library}{The base library to use as a starting point for the custom library. Default is "default_library". If base_library = "none", the function starts with an empty library.}

\item{custom_fd_list}{A list of custom feeds created using \code{\link[dairynasem]{create_custom_fd}}.}
}
\value{
A dataframe representing the custom library with appended custom feeds.
}
\description{
Create a custom library by appending custom feeds to the default library. The user can specify whether to use the default_library as the base library, or to specify a different library. To use only custom feeds, the user can also specify "none" as the base library.
}
\examples{

# Create a custom library using the default library as a base and appending custom feeds.

Custom_alfalfa_silage <- create_custom_fd("Legume silage, mid-maturity",
"Custom_alfalfa_silage",
list("Fd_DM" = 45, "Fd_CP" = 28))

Custom_corn_silage <- create_custom_fd("Corn silage, typical",
"Custom_corn_silage",
list("Fd_DM" = 38, "Fd_CP" = 8))

my_custom_library <- create_custom_library(base_library = "default_library",
list(Custom_alfalfa_silage, Custom_corn_silage))

}
