% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dairynasem-package.R
\docType{package}
\name{dairynasem-package}
\alias{dairynasem}
\alias{dairynasem-package}
\title{dairynasem: Wrapper functions to run the NASEM (2021) dairy nutritional model}
\description{
This package is designed for dairy nutrition researchers to efficiently run the NASEM 2021 dairy nutritional model. This package simplifies the process of running simulations on multiple animal and diet inputs.
}
\keyword{internal}
