% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_library}
\alias{default_library}
\title{default_library}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 304 rows and 87 columns.
}
\usage{
default_library
}
\description{
default_library
}
\references{
Nutrient Requirements of Dairy Cattle: Eighth Revised Edition. (2021). National Academies Press. \url{https://doi.org/10.17226/25806}
}
\author{
NASEM Dairy Model, imported by MaryGrace Erickson \email{eric3085@umn.edu}
}
\keyword{datasets}
