% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diet_comp_ex_1}
\alias{diet_comp_ex_1}
\title{diet_comp_ex_1}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.
}
\usage{
diet_comp_ex_1
}
\description{
diet_comp_ex_1
}
\references{
Nutrient Requirements of Dairy Cattle: Eighth Revised Edition. (2021). National Academies Press. \url{https://doi.org/10.17226/25806}
}
\author{
MaryGrace Erickson \email{eric3085@umn.edu}
}
\keyword{datasets}
