% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_result.R
\name{dn_export}
\alias{dn_export}
\title{Export text or spreadsheet results}
\usage{
dn_export(dn_out, filetype = "txt")
}
\arguments{
\item{dn_out}{The output of a Dairy NASEM simulation given by dn_run.}

\item{filetype}{A character string specifying the type of file to export. Valid options are "txt" (default) and "xlsx".}
}
\value{
This function does not have a direct return value. It creates and exports .txt or .xlsx files with the simulation results.
}
\description{
This exports a text or xlsx file from a dn_out object produced by dn_run.
}
\details{
Writes a .txt or .xlsx file to the working directory with simulation results for each element in a dn_out nested list (produced by \code{\link[dairynasem]{dn_run}}). Output is named with the DietID and ScenarioID.
}
\examples{

# Run simulations

data(diet_comp_ex_1)
data(scenario_ex_1)
dn_out = dn_run(diet_comp = diet_comp_ex_1, scenario = scenario_ex_1)


# Export the results to .txt files (default):

dn_export(dn_out)

# Export the results to .xlsx files:

dn_export(dn_out, filetype = "xlsx")

}
