% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_result.R
\name{dn_extract}
\alias{dn_extract}
\title{Extract some part of the simulation results}
\usage{
dn_extract(dn_out, which_output)
}
\arguments{
\item{dn_out}{the output of a Dairy NASEM simulation given by dn_run.}

\item{which_output}{which_output specifies which part of the NASEM results should be returned as a dataframe.}
}
\value{
A dataframe with selected results extracted from an object produced by dn_run.
}
\description{
This extracts an element of a dn_out object returned by dn_run, for all simulations.
}
\details{
which_output specifies which part of the NASEM results should be returned as a dataframe.
\itemize{
\item{Tbl1_1}{Physiological State/Management}
\item{Tbl1_2}{Entered Performance}
\item{Tbl1_3}{Predicted Production Variables}
\item{Tbl2_1}{Macronutrients}
\item{Tbl2_2}{Diet Ingredients}
\item{Tbl3_1}{Ingredient Macronutrient Contributions}
\item{Tbl4_1}{Energy Supply}
\item{Tbl4_2}{NEL and ME Requirements}
\item{Tbl4_3}{Nutrient Contributions to DE}
\item{Tbl5_1}{Fatty Acid Supply}
\item{Tbl6_1}{Animal Inputs}
\item{Tbl6_2}{Net Protein and Metabolizable Protein}
\item{Tbl6_3}{Predicted and Target Supply of Metabolizable Protein and Amino Acids}
\item{Tbl6_4}{EAA Partitioning using Predicted Milk NP, g/d}
\item{Tbl7_1}{Minerals}
\item{Tbl7_2}{Vitamin Supply and Requirements}
\item{Tbl7_3}{Water, Volatile Solids, and Methane}
\item{Tbl8_1}{Nitrogen and Mineral Excretion}
\item{Tbl8_2}{Ingredient Mineral Contributions}
}
}
\examples{

# Run simulations

data(diet_comp_ex_1)
data(scenario_ex_1)
dn_out = dn_run(diet_comp = diet_comp_ex_1, scenario = scenario_ex_1)


# Extract results related to dry matter intake

dmi_results <- dn_extract(dn_out, "dmi")


# Extract results for Table 4-1

table_4_1_results <- dn_extract(dn_out, "Tbl4_1")
}
