% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_report.R
\name{dn_report}
\alias{dn_report}
\title{Generate an R Markdown report of results.}
\usage{
dn_report(dn_simulation, filename, author)
}
\arguments{
\item{dn_simulation}{A single simulation, i.e., one element of a dn_run list object.}

\item{filename}{A character string specifying the filename.}

\item{author}{The name to add to the top of the R Markdown report.}
}
\value{
an Rmd file
}
\description{
This function creates an R Markdown report from the results of a dairynasem simulation. Only one element (i.e., one diet and scenario combination) of a dn_run object can be viewed in a report.
}
\examples{

# Run simulations

data(diet_comp_ex_1)
data(scenario_ex_1)
dn_out = dn_run(diet_comp = diet_comp_ex_1, scenario = scenario_ex_1)

# Generate a report

dn_report(dn_out[[1]], "file2", "My Name")

}
