% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_run.R
\name{dn_run}
\alias{dn_run}
\title{Run the NASEM (2021) model}
\usage{
dn_run(
  diet_comp,
  scenario,
  feed_library = "default_library",
  eff = "default_eff",
  infus = "no_infusions",
  diet_scenario_combos = NA
)
}
\arguments{
\item{diet_comp}{a dataframe with DietID, Fd_Name, and Fd_Amount_DM}

\item{scenario}{a dataframe with ScenarioID, and details of the scenario, e.g., An_Breed}

\item{feed_library}{a dataframe, default is default_library which includes NASEM (2021) values}

\item{eff}{either the character string "default_eff" or a vector of efficiencies. The default sets efficiencies to the NASEM (2021) values.}

\item{infus}{either the character string "no_infusions" or a vector of infusion data. The default is no_infusions.}

\item{diet_scenario_combos}{a character vector specifying which combinations of diet and scenario desired. If none are specified, the function returns all possible combinations of diets and scenarios.}
}
\value{
A list containing the results of Dairy NASEM simulations for each combination of diet and scenario.
}
\description{
This function runs the NASEM (2021) model. It produces all possible combinations of diets and scenarios to generate a set of simulations, unless a subset of diet_scenario_combos is specified to limit the results.
}
\examples{

data(diet_comp_ex_1)
data(scenario_ex_1)
dn_out = dn_run(diet_comp = diet_comp_ex_1, scenario = scenario_ex_1)

}
