% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_infusions}
\alias{no_infusions}
\title{no_infusions}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 34 columns.
}
\usage{
no_infusions
}
\description{
no_infusions
}
\references{
Nutrient Requirements of Dairy Cattle: Eighth Revised Edition. (2021). National Academies Press. \url{https://doi.org/10.17226/25806}
}
\author{
NASEM Dairy Model, imported by MaryGrace Erickson \email{eric3085@umn.edu}
}
\keyword{datasets}
