% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_ex_1}
\alias{scenario_ex_1}
\title{scenario_ex_1}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 38 columns.
}
\usage{
scenario_ex_1
}
\description{
scenario_ex_1
}
\references{
Nutrient Requirements of Dairy Cattle: Eighth Revised Edition. (2021). National Academies Press. \url{https://doi.org/10.17226/25806}
}
\author{
MaryGrace Erickson \email{eric3085@umn.edu}
}
\keyword{datasets}
